1) Добавлена история просмотров 2) Множественные исправления в связи с внедрением нового функционала 3) Исправления ошибок связанные с внедрением нового функционала

+++ app/src/main/java/ru/garretech/garred/doramatv/database/AppDataSource.kt
@@ -1,16 +1,15 @@
 package ru.garretech.garred.doramatv.database
 
 import android.content.Context
-import android.graphics.Bitmap
 import android.util.Log
-
-import java.util.ArrayList
+import io.reactivex.Completable
 
 import io.reactivex.Observable
 import io.reactivex.Single
 import ru.garretech.garred.doramatv.model.Favorites
+import ru.garretech.garred.doramatv.model.History
 import ru.garretech.garred.doramatv.model.Movie
-import ru.garretech.garred.doramatv.tools.SiteWorker
+import kotlin.collections.ArrayList
 
 class AppDataSource(context: Context) {
     private val appDatabase : AppDatabase by lazy { AppDatabase.getInstance(context)!! }
@@ -19,6 +18,9 @@ class AppDataSource(context: Context) {
     }
     private val favoritesDAO: FavoritesDAO by lazy {
         appDatabase.favoritesDAO()
+    }
+    private val historyDAO: HistoryDAO by lazy {
+        appDatabase.historyDAO()
     }
 
     val allMovies: List<Movie>
@@ -32,22 +34,45 @@ class AppDataSource(context: Context) {
 
             for (favorite in favorites) {
                 val movie = movieDAO.getMovie(favorite.movieURL)
-                list.add(movie)
+                list.add(movie!!)
             }
             return Observable.fromArray(list)
         }
 
-    init {
+    val listOfHistoryObservable : Observable<List<Movie>>
+        get() {
+            val historyList = historyDAO.allHistory
+            val movieList = ArrayList<Movie>()
 
-    }
+            for (history in historyList) {
+                val movie = movieDAO.getMovie(history.movieURL)
+                if (movie != null)
+                    movieList.add(movie)
+            }
+            return Observable.fromArray(movieList)
+        }
+
+    val listOfHistorySingle =
+            Single.create<List<History>> {
+                val list = historyDAO.allHistory
+
+                it.onSuccess(list)
+            }
+
 
     fun addMovie(movie: Movie) {
         movieDAO.addMovie(movie)
     }
 
-    fun getMovie(URL: String): Movie {
-        return movieDAO.getMovie(URL)
-    }
+    fun getMovie(URL: String) =
+            Single.create<Movie> {
+                val movie = movieDAO.getMovie(URL)
+
+                if (movie != null)
+                    it.onSuccess(movie)
+                else
+                    it.onError(NullPointerException())
+            }
 
     fun isFavorite(URL: String): Boolean {
         val favorites = favoritesDAO.getFavoriteByURL(URL)
@@ -57,15 +82,57 @@ class AppDataSource(context: Context) {
     fun addFavorites(movie: Movie) {
         movieDAO.addMovie(movie)
         Log.d("Database", "Movie " + movie.url + " added")
-        val favorite = Favorites()
-        favorite.movieURL = movie.url
+        val favorite = Favorites(movie.url)
         favoritesDAO.addFavorites(favorite)
     }
 
 
     fun deleteFavorites(movie: Movie) {
         val favorites = favoritesDAO.getFavoriteByURL(movie.url)
-        favoritesDAO.deleteFavorites(favorites)
+        if (favorites != null)
+            favoritesDAO.deleteFavorites(favorites)
     }
 
+    fun getHistory(movie: Movie) =
+        Single.create<History> {
+            val history = historyDAO.getHistoryByURL(movie.url)
+
+            if (history == null)
+                it.onSuccess(History((movie.url)))
+            else
+                it.onSuccess(history)
+
+        }
+
+    fun addHistory(history: History) =
+            Completable.fromCallable {
+                historyDAO.addHistory(history)
+            }
+
+    fun addHistory(movie: Movie)  {
+                var history = historyDAO.getHistoryByURL(movie.url)
+
+                if (history == null) {
+                    history = History(movie.url).also { it.series = HashMap() }
+                    historyDAO.addHistory(history)
+                }
+
+            }
+
+
+    fun updateHistory(history: History) =
+            Completable.fromCallable {
+                historyDAO.updateHistory(history)
+            }
+
+
+    fun clearHistory() =
+            Completable.fromCallable {
+                historyDAO.clearHistory()
+            }
+
+    fun clearFavorites() =
+            Completable.fromCallable {
+                favoritesDAO.clearFavorites()
+            }
 }
